backend/requirements.txt
fastapi
uvicorn
python-multipart
python-vlc
ultralytics
pillow
